backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Ata <ataonat00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
todoist-api-python = "^2.1.7"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
